se agrego los service, crud, exception, repository de pedido, producto y empleado

+++ src/main/java/org/kaven/Cafeteria/persistence/crud/CrudEstudianteEntity.java
@@ -3,6 +3,6 @@ package org.kaven.Cafeteria.persistence.crud;
 import org.kaven.Cafeteria.persistence.entity.EstudianteEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CrudEstudianteEntity extends CrudRepository<EstudianteEntity, Integer> {
+public interface CrudEstudianteEntity extends CrudRepository<EstudianteEntity, Long> {
     EstudianteEntity findFirstByNombre(String nombre);
 }
